feat(notification): add rate limiting configuration and enhance dispatcher
This update introduces a new RateLimitingConfig structure to manage rate limiting settings for push notifications. The configuration includes options for enabling rate limiting, setting requests per minute, and defining burst size.

Key changes:
- Added RateLimitingConfig to config.go and corresponding YAML configuration in config.yaml.
- Updated defaults.go to set default values for rate limiting.
- Enhanced the push dispatcher to utilize the new rate limiting features, ensuring that notifications do not overwhelm external APIs.
- Refactored push dispatcher tests to accommodate the new enhanced provider structure.

This implementation lays the groundwork for future integration of rate limiting functionality in push notifications, improving overall system reliability and performance.
